requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-text-splitters==0.0.1
langchain-mistralai==0.0.5
faiss-cpu==1.7.4
pypdf==3.17.4
numpy==1.24.3
gunicorn==21.2.0
tensorflow-cpu==2.15.0
langgraph==0.0.20
typing-extensions==4.9.0